refactor: move executeCommandAction to its own file
